feat: add repository and database

+++ CurrencyCompanion/Code/Entity/CurrencyRate.swift
@@ -10,5 +10,5 @@ import Foundation
 struct CurrencyRate {
   let baseCurrency: String
   let targetCurrency: String
-  let rate: Double
+  let rate: Float
 }
